Refactor tests and fix type issue
